First cut at the automatic update checking infrastructure.  So far, there's an XML document type for describing the available components and all versions thereof, and a class that can fully parse it, creating a graph of objects that describe products and versions.
The major thing that's not yet addressed is examining the local environment to determine which versions of the various components are available.  I think this could be a simple API of registerProductVersion(ProductVersion) on startup (or upon loading a particular component), and then a getLatestRegisteredVersion(Product) to check what we have, and compare it against the available updates.  We could also warn the user when starting a product and a newer version of the same product has been registered on the local system in the past.