-only show the editors panel when there are multiple robots -key bindings 1, 2, 3 ... for switching between circuits to edit -animated transtitions between circuit editors (actually just an effect on the selection frame.. the zoom effect was too slow) -label the editors in the editors panel -editors in the editors panel work like radio buttons, switching the current editor -title section with robot names in editors panel -make the hotkeys for the circuit editor work again
Note about the zoom effect: This was initially intended as the effect for switching between editors in the EditorsPanel (the thumbnail would zoom up to the main editor position beside the playfield), but it was too slow when it had to repaint the entire game UI on every frame.  I'm checking it in because it's generally useful and it will likely find a purpose later on.